memberRoom refactoring

+++ src/main/java/clofi/codeython/room/repository/jpa/RoomMemberJpaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RoomMemberJpaRepository extends JpaRepository<RoomMember, Integer> {
+public interface RoomMemberJpaRepository extends JpaRepository<RoomMember, Long> {
     List<RoomMember> findAllByRoom(Room room);
 
     List<RoomMember> findAllByRoomRoomNo(Long roomNo);
